modified outputs

+++ ns-3.29/src/mix-autonomy/model/hdv.h
@@ -20,7 +20,7 @@ public:
 
     enum Status {Middle, Outter};
     //static int GetHumanStatus (Ptr<Node> node);
-    uint8_t GetStatus ();
+    int GetStatus ();
     void SetTDMAEnable (bool val);
 
     void AddToMiddle(); //如果收到内核层包，更新状态信息
